+++ src/main/java/tornadofx/Workspace.kt
@@ -139,9 +139,11 @@ open class Workspace(title: String = "Workspace", navigationMode: NavigationMode
                         addClass("icon-only")
                         backButton = this
                         graphic = label { addClass("icon", "back") }
-                        setOnAction {
-                            viewPos.set(viewPos.get() - 1)
-                            dock(viewStack[viewPos.get()], false)
+                        action {
+                            if (dockedComponent?.onNavigateBack() ?: true) {
+                                viewPos.set(viewPos.get() - 1)
+                                dock(viewStack[viewPos.get()], false)
+                            }
                         }
                         disableProperty().bind(booleanBinding(viewPos, viewStack) { value < 1 })
                     }
@@ -149,9 +151,11 @@ open class Workspace(title: String = "Workspace", navigationMode: NavigationMode
                         addClass("icon-only")
                         forwardButton = this
                         graphic = label { addClass("icon", "forward") }
-                        setOnAction {
-                            viewPos.set(viewPos.get() + 1)
-                            dock(viewStack[viewPos.get()], false)
+                        action {
+                            if (dockedComponent?.onNavigateForward() ?: true) {
+                                viewPos.set(viewPos.get() + 1)
+                                dock(viewStack[viewPos.get()], false)
+                            }
                         }
                         disableProperty().bind(booleanBinding(viewPos, viewStack) { value == viewStack.size - 1 })
                     }
@@ -162,7 +166,7 @@ open class Workspace(title: String = "Workspace", navigationMode: NavigationMode
                         graphic = label {
                             addClass("icon", "refresh")
                         }
-                        setOnAction {
+                        action {
                             onRefresh()
                         }
                     }
@@ -171,7 +175,7 @@ open class Workspace(title: String = "Workspace", navigationMode: NavigationMode
                         saveButton = this
                         isDisable = true
                         graphic = label { addClass("icon", "save") }
-                        setOnAction {
+                        action {
                             onSave()
                         }
                     }